fix: make auth form background transparent to show layout gradient
