--- src/components_ui/LoginPage.tsx
@@ -1,17 +1,0 @@
-import { Box } from "@chakra-ui/react";
-import LoginForm from "./LoginForm";
-
-export default function LoginPage() {
-	return (
-		<Box
-			flex="1"
-			display="flex"
-			alignItems="center"
-			justifyContent="center"
-			bg="gray.50"
-			p={4}
-		>
-			<LoginForm />
-		</Box>
-	);
-}
