+++ src/main.tsx
@@ -3,6 +3,7 @@ import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "./components/ui/provider.tsx";
+import { Box } from "@chakra-ui/react";
 
 const rootElement = document.getElementById("root");
 
@@ -11,7 +12,13 @@ if (rootElement) {
 		<StrictMode>
 			<Provider>
 				<BrowserRouter>
-					<App />
+					<Box
+						minH="100vh"
+						w="full"
+						bgGradient="linear(to-br, green.50, green.100)"
+					>
+						<App />
+					</Box>
 				</BrowserRouter>
 			</Provider>
 		</StrictMode>,
