+++ src/modules/auth/formsAuth/LoginForm.tsx
@@ -1,11 +1,11 @@
 import { Stack, Button, Box, VStack } from "@chakra-ui/react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { LoginSchema } from "./types/types";
+import { LoginSchema } from "../../types/types";
 import type { z } from "zod";
 import FormInput from "./AuthForm";
 import { useNavigate } from "react-router-dom";
-import { useUserStore } from "../store/UserStore";
+import { useUserStore } from "../../../store/UserStore";
 
 type LoginFormFields = z.infer<typeof LoginSchema>;
