+++ src/modules/auth/pagesAuth/LoginPage.tsx
@@ -1,0 +1,10 @@
+import { Box } from "@chakra-ui/react";
+import LoginForm from "../formsAuth/LoginForm";
+
+export default function LoginPage() {
+	return (
+		<Box w="full" display="flex" justifyContent="center" bg="transparent">
+			<LoginForm />
+		</Box>
+	);
+}
